+++ src/main/java/org/jboss/aerogear/android/core/ConfigurationProvider.java
@@ -14,21 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.android.impl.reflection;
+package org.jboss.aerogear.android.core;
 
-/**
- * Thrown when a particular field cannot be found.
- */
-public class FieldNotFoundException extends RuntimeException {
+public interface ConfigurationProvider<CFG extends Config<CFG>> {
 
-    /**
-     * Constructs with class and fieldName cannot be found.
-     *
-     * @param klass Class where field cannot be found.
-     * @param fieldName field cannot be found.
-     */
-    public FieldNotFoundException(Class klass, String fieldName) {
-        super("Cannot find field " + fieldName + " on " + klass.getSimpleName());
-    }
+    CFG newConfiguration();
 
 }
