+++ src/main/java/org/jboss/aerogear/android/core/ReadFilter.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.android;
+package org.jboss.aerogear.android.core;
 
 import android.util.Log;
 import org.json.JSONException;
