+++ src/main/java/org/jboss/aerogear/android/core/reflection/Property.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.android.impl.reflection;
+package org.jboss.aerogear.android.core.reflection;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
