+++ src/main/java/org/jboss/aerogear/android/core/reflection/PropertyNotFoundException.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.android.impl.reflection;
+package org.jboss.aerogear.android.core.reflection;
 
 /**
  * Thrown when a property cannot be found.
