+++ src/main/java/org/jboss/aerogear/android/core/reflection/Scan.java
@@ -14,9 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.android.impl.reflection;
+package org.jboss.aerogear.android.core.reflection;
 
-import org.jboss.aerogear.android.RecordId;
+import org.jboss.aerogear.android.core.RecordId;
 
 import java.lang.reflect.Field;
